Create README.md
The store keeps a list of customers so that when something is purchased from the store, it is 
recorded against that customer’s name. Only the customer’s name is required.
Only two items are currently sold in the store currently – books and magazines, and only the name 
of the item is required. Clearly, a list of such items is needed, and the user should be able to add 
items to the list. When an item is added, the application should automatically make a backup in