Add files via upload
The store keeps a list of customers so that when something is purchased from the store, it is 
recorded against that customer’s name. Only the customer’s name is required.
Only two items are currently sold in the store currently – books and magazines, and only the name of the item is required. Clearly, a list of such items is needed, and the user should be able to add items to the list. When an item is added, the application should automatically make a backup in memory in case something goes wrong when the application is being used. Provision should thus be made to restore this list when necessary  When a customer purchases items, a transaction is recorded. The date/time of the purchase is noted, as is the name, type, and quantity of each item purchased as part of the transaction. Use an appropriate widget to indicate which items have already been added as part of this transaction.